M9c: gl4es web build verified on stock emscripten; 1984 libimage LP64 fix.
- libimage: RLE rowstart/rowsize tables declared long* but allocated and
  file-read as int32 - on LP64 this stomps the heap and yields garbage
  row seeks (black textures on ANY backend; IRIX/wasm ILP32 were immune).
  Found with guard malloc. Fields are now unsigned int*/int*.
- SoTexture2: full fiopen/iclose prototypes (K&R 2-arg decl traps in wasm).
- Web via gl4es verified in-browser (maze + slotcar title and race) on
  unpatched emscripten with FULL_ES2; textured cube pixel-identical to
  desktop GL natively.
- README: gl4es backend build instructions; ARCHITECTURE.md: M9 row;
  .gitignore: build-gl4es*/ dirs (untracked accidentally committed
  binaries from M9a).

+++ inventor-sdl2-gles2/libimage/image.h
@@ -108,8 +108,13 @@ typedef struct {
     unsigned short	*tmpbuf;
     unsigned long	offset;
     unsigned long	rleend;		/* for rle images */
-    unsigned long	*rowstart;	/* for rle images */
-    long		*rowsize;	/* for rle images */
+    /* The RLE offset/length tables are 32-bit in the file format and were
+     * allocated/read as N*sizeof(int) bytes; on IRIX (ILP32) 'long' matched.
+     * On LP64 platforms 'long' is 8 bytes: indexing 8-byte entries in a
+     * 4-byte-entry allocation both reads garbage offsets and writes past
+     * the heap block (initialization loop) -> corrupted textures + heap. */
+    unsigned int	*rowstart;	/* for rle images */
+    int			*rowsize;	/* for rle images */
 } IMAGE;
 
 IMAGE *icreate();
